Calculate function added.
the calculator now works. it takes a simple equation with 2 operators and an operand or it takes an equation with multiple operands. in the later case it calculates from left to right rather than with proper order of operations. for this simple version it's assumed the user it doing the order of operations.

There has been no error checking or messaging yet.